feat: :star2: add exercise 4.18 - 4.23

+++ part4/bloglist/src/modules/blog/controller.js
@@ -1,6 +1,7 @@
 const Router = require('express').Router
 const service = require('./service')
 const { ServerError } = require('../../utils/error')
+const middleware = require('../../utils/middleware')
 
 const router = Router()
 
@@ -8,13 +9,13 @@ router.get('/', async (_req, res) => {
   res.json(await service.get())
 })
 
-router.post('/', async (req, res) => {
-  const { title, author, url, userId, likes } = req.body
+router.post('/', middleware.authGuard, async (req, res) => {
+  const { title, author, url, likes } = req.body
+  const user = req.user
 
   if (title == null && url == null) throw new ServerError('Missing post title and URL', 400)
-  if (userId == null) throw new ServerError('Missing userId', 400)
 
-  const newEntry = await service.create({ title, author, url, userId, likes: likes ?? 0 })
+  const newEntry = await service.create({ title, author, url, userId: user.id, likes: likes ?? 0 })
 
   res.status(201).json(newEntry)
 })
@@ -35,10 +36,11 @@ router.put('/:id', async (req, res) => {
   res.json(entry)
 })
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', middleware.authGuard, async (req, res) => {
   const { id } = req.params
+  const user = req.user
 
-  await service.remove({ id })
+  await service.remove({ id, userId: user.id })
 
   res.status(204).end()
 })
